better form structure
