--- components/ProposalForms/FormFactory.js
@@ -1,9 +1,0 @@
-import keys from 'components/ProposalForms/keys'
-
-export default class FormFactory {
-  static createForm(value) {
-    return { ...new value.className(value.key) }
-  }
-
-  static Keys = keys
-}
