--- components/ProposalForms/GenericForm.js
@@ -1,8 +1,0 @@
-export default class GenericForm {
-  constructor(key) {
-    this.key = key;
-    this.title = '';
-    this.schema = {};
-    this.processData = () => {};
-  }
-}
