+++ styles/globals.css
@@ -162,5 +162,5 @@ nav .text-2xl{
 
 .rjsf textarea {
     font-family: ui-monospace,SFMono-Regular,Menlo,Monaco,Consolas,Liberation Mono,Courier New,monospace;
-    height: 20rem;   
+    height: 6rem;   
 }
