+++ types/global.d.ts
@@ -1,0 +1,4 @@
+interface Window {
+  GroupAddress: string
+  MultisigAddress: string
+}
